Reusing enentity creator
